+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## HEAD (Unreleased)
+_(none)_
+
+--------------------
+
+## 0.5.4 (2016-12-28)
+
+* Feature: 去除了Plumber\Logger类，使用Monolog类替代。
+* Feature: 使用Monolog\ErrorHandler捕获程序错误，并记录日志。
+* Feature: 去除了`output_path`的配置，output输出的内容，合并到`log_path`所在文件。
+
 ## v0.5.3 (2016-07-26)
 
 * Fixed: 当retry job时，向job body注入重试次数的变量名由`retry`改为`__retry`。